+++ Classes/UVBaseSuggestionViewController.h
@@ -41,7 +41,6 @@
 
 - (id)initWithTitle:(NSString *)theTitle;
 - (void)pushCategorySelectView;
-- (UITextField *)customizeTextFieldCell:(UITableViewCell *)cell label:(NSString *)label placeholder:(NSString *)placeholder;
 - (void)reloadCategoryTable;
 
 @end