+++ Classes/UVUser.m
@@ -155,7 +155,7 @@
         self.supportedSuggestions = [NSMutableArray array];
         
         self.visibleForumsDict = [self objectOrNilForDict:dict key:@"visible_forums"];
-        if ([UVSession currentSession].clientConfig.forum)
+        if ([UVSession currentSession].forum)
           [self updateVotesRemaining];
     }
     return self;
@@ -163,7 +163,7 @@
 
 - (void)updateVotesRemaining {
     for (NSDictionary *forum in self.visibleForumsDict) {
-        if ([(NSNumber *)[forum valueForKey:@"id"] integerValue] == [UVSession currentSession].clientConfig.forum.forumId) {
+        if ([(NSNumber *)[forum valueForKey:@"id"] integerValue] == [UVSession currentSession].forum.forumId) {
             NSDictionary *activity = [self objectOrNilForDict:forum key:@"forum_activity"];
             self.votesRemaining = [(NSNumber *)[activity valueForKey:@"votes_available"] integerValue];
         }